+++ graphs/src/cz/nkp/urnnbn/client/charts/StatisticsServiceAsync.java
@@ -1,4 +1,4 @@
-package cz.nkp.urnnbn.client;
+package cz.nkp.urnnbn.client.charts;
 
 import java.io.Serializable;
 import java.util.HashMap;
@@ -8,8 +8,8 @@ import java.util.Set;
 
 import com.google.gwt.user.client.rpc.AsyncCallback;
 
-import cz.nkp.urnnbn.shared.Registrar;
-import cz.nkp.urnnbn.shared.Statistic;
+import cz.nkp.urnnbn.shared.charts.Registrar;
+import cz.nkp.urnnbn.shared.charts.Statistic;
 
 public interface StatisticsServiceAsync {
