+++ graphs/src/cz/nkp/urnnbn/client/charts/widgets/IntegerSelectionHandler.java
@@ -1,4 +1,4 @@
-package cz.nkp.urnnbn.client.widgets;
+package cz.nkp.urnnbn.client.charts.widgets;
 
 public interface IntegerSelectionHandler {
 	public void onSelected(Integer value);
