+++ graphs/src/cz/nkp/urnnbn/client/charts/widgets/RegistrarWithStatistic.java
@@ -1,4 +1,4 @@
-package cz.nkp.urnnbn.client.widgets;
+package cz.nkp.urnnbn.client.charts.widgets;
 
 public class RegistrarWithStatistic implements Comparable<RegistrarWithStatistic> {
 	private final String code;
