+++ graphs/src/cz/nkp/urnnbn/client/charts/widgets/StringSelectionHandler.java
@@ -1,4 +1,4 @@
-package cz.nkp.urnnbn.client.widgets;
+package cz.nkp.urnnbn.client.charts.widgets;
 
 public interface StringSelectionHandler {
